backend/requirements.txt
# Tatva.ai Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
pydantic==2.5.0
python-dotenv==1.0.0

# Optional: For IndicWav2Vec support (future)
# transformers==4.35.0
# accelerate==0.24.0
